pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lagoon"
version = "0.1.0"
description = "Text-scoring library that maps input text to 207 semantic reefs using precomputed BM25 scores"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "msgpack>=1.0",
    "pyahocorasick>=2.0",
    "PyStemmer>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-benchmark",
]

[tool.hatch.build.targets.wheel]
packages = ["src/lagoon"]
